requirements.txt
pyOpenSSL==20.0.1
gunicorn==20.0.4
Flask_Bcrypt==0.7.1
Flask_Uploads==0.2.1
Werkzeug==0.15.4
Flask_SQLAlchemy==2.4.0
Flask_Mail==0.9.1
Flask_WTF==0.14.2
itsdangerous==1.1.0
pandas==0.24.2
Flask_Toastr==0.5.2
Flask==1.1.1
Flask_Login==0.4.1
firebase_admin==4.5.1
firebase==3.0.1
numpy==1.16.4
arrow==0.17.0
Flask_DebugToolbar==0.10.1
Flask_Simple_GeoIP==0.2.4
WTForms_SQLAlchemy==0.1
requests==2.22.0
urllib3==1.24.2
WTForms==2.2.1
certifi==2020.12.5
Pillow==8.1.2
python-jwt==3.3.0
gcloud == 0.18.3
sseclient == 0.0.27
crypto == 1.4.1
pycryptodome==3.4.3
requests-toolbelt == 0.9.1
plotly
chart_studio
